Ajustes en los mantenimientos de usuario y doctor

+++ OrientalMedical.WebService/OrientalMedical.Services/Validations/PersonalValidations.cs
@@ -32,7 +32,7 @@ namespace OrientalMedical.Services.Validations
         {
             try
             {
-                int.Parse(celula);
+                long.Parse(celula);
             }
             catch (Exception)
             {
